+++ packages/spotlight/CHANGELOG.md
@@ -10,6 +10,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
+- `spotlight` to handle non-5-way keys correctly to focus on next 5-way keys
+
 ## [1.12.1] - 2017-11-07
 
 No significant changes.